Move components to sub folder
