+++ src/components/layout/index.ts
@@ -1,3 +1,4 @@
 export * from './Footer';
 export * from './Layout';
 export * from './Navbar';
+export * from './Content';
