+++ src/components/seo/Seo.tsx
@@ -7,7 +7,7 @@
 
 import React from "react";
 import { Helmet } from "react-helmet";
-import { useSiteMetadata } from '../hooks/useSiteMetadata';
+import { useSiteMetadata } from '../../hooks/useSiteMetadata';
 
 interface SeoProps {
   description?: string;
