+++ src/components/theme-toggle/ThemeToggle.tsx
@@ -1,8 +1,8 @@
-import { useIsomorphicLayoutEffect } from '../hooks/useIsomorphicLayoutEffect'
+import { useIsomorphicLayoutEffect } from '../../hooks/useIsomorphicLayoutEffect'
 import { Listbox } from '@headlessui/react'
 import React, { Fragment, useEffect, useRef, useState } from 'react'
-import { Pc } from './icons/svg/device/Icons'
-import { Moon, Sun } from './icons/svg/weather/Icons'
+import { Pc } from '../icons/svg/device/Icons'
+import { Moon, Sun } from '../icons/svg/weather/Icons'
 
 function update() {
     if (
