+++ src/pages/404.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Layout } from '../components/layout';
-import { Link } from 'gatsby';
 import { ButtonInternalLink } from '../components/button';
 
 const NotFoundPage = () => (
