+++ src/templates/index-page.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { graphql, PageProps } from "gatsby";
 import { Layout } from '../components/layout';
 import { IndexPageTemplate } from "./index-page-template";
-import { Seo } from "../components/Seo";
+import { Seo } from "../components/seo/Seo";
 import { HomePageQuery } from "../types/graphql-queries";
 
 const IndexPage = ({ data: { markdownRemark: { frontmatter } } }: PageProps<HomePageQuery>) => {
